Se soluciona el guardado de videojuegos

+++ src/main/java/com/martincarrion/videojuegosapp/entities/Videojuego.java
@@ -21,7 +21,7 @@ public class Videojuego {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private long id;
 
     @Column(name = "titulo")
     private String titulo;
@@ -33,7 +33,7 @@ public class Videojuego {
     private String imagen;
 
     @Column(name = "precio")
-    private Double precio;
+    private float precio;
 
     @Column(name = "stock")
     private short stock;
